refactor(hamilton/tcp): replace status_exception with HoiError module Move HOI error parsing (parse_hamilton_error_*) from messages into hoi_error. Pass HOI entries and per-entry exceptions into ChannelizedError for callers.
