+++ pylabrobot/hamilton/tcp/__init__.py
@@ -1,7 +1,12 @@
 """Canonical v1 Hamilton TCP namespace."""
 
 from pylabrobot.hamilton.tcp.client import HamiltonTCPClient
-from pylabrobot.hamilton.tcp.status_exception import HamiltonStatusException
+from pylabrobot.hamilton.tcp.hoi_error import (
+  HoiError,
+  parse_hamilton_error_entries,
+  parse_hamilton_error_entry,
+  parse_hamilton_error_params,
+)
 from pylabrobot.hamilton.tcp.interface_bundle import InterfacePathSpec, resolve_interface_path_specs
 from pylabrobot.hamilton.tcp.commands import TCPCommand
 from pylabrobot.hamilton.tcp.introspection import (
@@ -51,7 +56,10 @@ __all__ = [
   "flatten_firmware_tree",
   "HAMILTON_PROTOCOL_VERSION_MAJOR",
   "HAMILTON_PROTOCOL_VERSION_MINOR",
-  "HamiltonStatusException",
+  "HoiError",
+  "parse_hamilton_error_entries",
+  "parse_hamilton_error_entry",
+  "parse_hamilton_error_params",
   "TCPCommand",
   "HamiltonTCPClient",
   "HamiltonDataType",
